# swagger/swagger_output.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "oneM2M API",
        "description": "oneM2M API 1.0",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Default",
            "variables": {}
        }
    ],
    "paths": {
...
        "/api/onem2m/devices/group/{ae}": {
            "post": {
                "tags": [
                    "Group"
                ],
                "description": "Group 리소스 등록",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_Group_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_Group_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GRP_Create"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "GRP 리소스 등록",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/onem2m/devices/group/{ae}/{grp}": {
            "get": {
                "tags": [
                    "Group"
                ],
                "description": "Application Entity 리소스 하위 Group 리소스 조회",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_Group_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_Group_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "grp",
                        "in": "path",
                        "description": "GRP resourceName, GRP 리소스의 이름",
                        "example": "Test_GRP_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "AE 하위 GRP 리소스 조회",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Group"
                ],
                "description": "Group 리소스 업데이트",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_Group_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_Group_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "grp",
                        "in": "path",
                        "description": "GRP resourceName, GRP 리소스의 이름",
                        "example": "Test_GRP_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GRP_Update"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "GRP 리소스 업데이트",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Group"
                ],
                "description": "Group 리소스 삭제",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_Group_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_Group_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "grp",
                        "in": "path",
                        "description": "GRP resourceName, GRP 리소스의 이름",
                        "example": "Test_GRP_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "GRP 리소스 삭제",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/onem2m/devices/group/{ae}/{grp}/data": {
            "get": {
                "tags": [
                    "Group"
                ],
                "description": "Group 리소스 멤버 조회",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_Group_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_Group_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "grp",
                        "in": "path",
                        "description": "GRP resourceName, GRP 리소스의 이름",
                        "example": "Test_GRP_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "GRP 리소스 멤버 조회",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Group"
                ],
                "description": "Group 리소스 멤버 데이터 제어",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_Group_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_Group_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "grp",
                        "in": "path",
                        "description": "GRP resourceName, GRP 리소스의 이름",
                        "example": "Test_GRP_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GRP_Member_Control"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "GRP 리소스 멤버 데이터 제어",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GRP_Create": {
                "type": "object",
                "properties": {
                    "rn": {
                        "type": "string",
                        "description": "resourceName, 리소스의 이름",
                        "example": "Test_GRP_01"
                    },
                    "mid": {
                        "type": "array",
                        "description": "memberIDs, Group 리소스의 멤버 ID",
                        "example": [
                            "/Mobius/Test_AE_01/Test_CNT_01",
                            "/Mobius/Test_AE_02/Test_CNT_02"
                        ]
                    },
                    "mnm": {
                        "type": "integer",
                        "description": "maxNrOfMembers, 최대 멤버 수",
                        "example": 10
                    }
                },
                "required": [
                    "rn",
                    "mid",
                    "mnm"
                ]
            },
            "GRP_Member_Control": {
                "type": "object",
                "properties": {
                    "con": {
                        "type": "string",
                        "description": "content, 컨텐츠 (센서 데이터, 위치 정보 등)",
                        "example": "off"
                    },
                    "lbl": {
                        "type": "array",
                        "description": "labels, 리소스의 Tag 정보",
                        "example": "Lamps_01"
                    }
                },
                "required": [
                    "con"
                ]
            },
            "GRP_Update": {
                "type": "object",
                "properties": {
                    "mnm": {
                        "type": "integer",
                        "description": "maxNrOfMembers, 최대 멤버 수",
                        "example": 15
                    }
                }
            },
...
        }
    }
}